refactor(suite): send form common thunks imports
This commit fixes imports of all the send form thunk moved  to suite-common in the previous commit. These two commits are related and should be reviewed together. Only reason why I decided to split them is to keep git history clean.
