+++ packages/suite/src/hooks/wallet/useCoinmarketRecomposeAndSign.ts
@@ -1,15 +1,13 @@
 import { useCallback } from 'react';
 import { useDispatch, useSelector, useTranslation } from 'src/hooks/suite';
-import {
-    signAndPushSendFormTransactionThunk,
-    composeSendFormTransactionThunk,
-} from 'src/actions/wallet/send/sendFormThunks';
+import { signAndPushSendFormTransactionThunk } from 'src/actions/wallet/send/sendFormThunks';
 
 import { notificationsActions } from '@suite-common/toast-notifications';
 import { DEFAULT_VALUES, DEFAULT_PAYMENT } from '@suite-common/wallet-constants';
 import { FormState, UseSendFormState } from 'src/types/wallet/sendForm';
 import { getFeeLevels } from '@suite-common/wallet-utils';
 import type { FormOptions, SelectedAccountLoaded } from '@suite-common/wallet-types';
+import { composeSendFormTransactionThunk } from '@suite-common/wallet-core';
 
 export const useCoinmarketRecomposeAndSign = () => {
     const { translationString } = useTranslation();
