+++ packages/suite/src/middlewares/wallet/walletMiddleware.ts
@@ -17,7 +17,7 @@ import { ROUTER } from 'src/actions/suite/constants';
 import { WALLET_SETTINGS } from 'src/actions/settings/constants';
 import * as selectedAccountActions from 'src/actions/wallet/selectedAccountActions';
 import { sendFormActions } from '@suite-common/wallet-core';
-import { convertSendFormDraftsThunk } from 'src/actions/wallet/send/sendFormThunks';
+import { convertSendFormDraftsThunk } from '@suite-common/wallet-core';
 import * as modalActions from 'src/actions/suite/modalActions';
 import * as receiveActions from 'src/actions/wallet/receiveActions';
 import * as cardanoStakingActions from 'src/actions/wallet/cardanoStakingActions';
