+++ suite-common/wallet-core/src/send/sendFormSolanaThunks.ts
@@ -10,7 +10,6 @@ import {
     PrecomposedLevels,
 } from '@suite-common/wallet-types';
 import { notificationsActions } from '@suite-common/toast-notifications';
-import { selectBlockchainBlockInfoBySymbol, selectDevice } from '@suite-common/wallet-core';
 import { createThunk } from '@suite-common/redux-utils';
 import {
     amountToSatoshi,
@@ -19,7 +18,6 @@ import {
     formatAmount,
     getExternalComposeOutput,
 } from '@suite-common/wallet-utils';
-
 import {
     getPubKeyFromAddress,
     buildTransferTransaction,
@@ -28,8 +26,10 @@ import {
     dummyPriorityFeesForFeeEstimation,
 } from '@suite-common/wallet-utils';
 
-import { MODULE_PREFIX } from './constants';
-import { ComposeTransactionThunkArguments, SignTransactionThunkArguments } from './types';
+import { selectBlockchainBlockInfoBySymbol } from '../blockchain/blockchainReducer';
+import { selectDevice } from '../device/deviceReducer';
+import { ComposeTransactionThunkArguments, SignTransactionThunkArguments } from './sendFormTypes';
+import { SEND_MODULE_PREFIX } from './sendFormConstants';
 
 const calculate = (
     availableBalance: string,
@@ -123,7 +123,7 @@ const fetchAccountOwnerAndTokenInfoForAddress = async (
 };
 
 export const composeSolanaSendFormTransactionThunk = createThunk(
-    `${MODULE_PREFIX}/composeSolanaSendFormTransactionThunk`,
+    `${SEND_MODULE_PREFIX}/composeSolanaSendFormTransactionThunk`,
     async ({ formValues, formState }: ComposeTransactionThunkArguments, { getState }) => {
         const { account, network, feeInfo } = formState;
         const composeOutputs = getExternalComposeOutput(formValues, account, network);
@@ -251,7 +251,7 @@ export const composeSolanaSendFormTransactionThunk = createThunk(
 );
 
 export const signSolanaSendFormTransactionThunk = createThunk(
-    `${MODULE_PREFIX}/signSolanaSendFormTransactionThunk`,
+    `${SEND_MODULE_PREFIX}/signSolanaSendFormTransactionThunk`,
     async (
         { formValues, transactionInfo, selectedAccount }: SignTransactionThunkArguments,
         { dispatch, getState, extra },
